use bytes

+++ skiplist/src/lib.rs
@@ -1,6 +1,6 @@
 mod arena;
 mod list;
 
-const MAX_HEIGHT: u32 = 20;
+const MAX_HEIGHT: usize = 20;
 
 pub use self::list::{IterRef, Skiplist};
